+++ backend/search_chroma.py
@@ -1,6 +1,6 @@
 import chromadb
 
-from hgjkl.vector_db_pipline import vectorize_text_segments
+from backend.vector_db_pipline import vectorize_text_segments
 
 
 def search_chroma(
